fix(validation): stop extreme exponents overflowing decimal bounds (#1765)
* fix(validation): compute integer digits in long to stop exponent overflow

* test(validation): cover extreme-exponent numeric overflow

* chore(api-surface): regenerate policy ledger

* fix(validation): cap request decimals with DecimalMax against exponent overflow

* test(validation): cover extreme exponents on request DTO decimal bounds

* fix(validation): backstop product and approval-policy amounts in services

* style(goals): wrap target precision check

* test(goals): assert target refusal message

* chore(api-surface): regenerate policy ledger

+++ backend/src/main/java/ooo/klae/connex/backend/services/RecordCreationTemplateValidator.java
@@ -371,7 +371,7 @@ public class RecordCreationTemplateValidator {
 
     private void validateNumberDefault(String fieldKey, BigDecimal value) {
         int fractionalDigits = Math.max(value.scale(), 0);
-        int integerDigits = Math.max(value.precision() - value.scale(), 0);
+        long integerDigits = Math.max((long) value.precision() - value.scale(), 0L);
         boolean invalid = "value".equals(fieldKey)
             ? value.signum() < 0 || fractionalDigits > 2 || integerDigits > 13
             : fractionalDigits > 4 || integerDigits > 16;
